added unit test for PDM support
