+++ tests/cli/test_cli.py
@@ -17,24 +17,6 @@ def dir_with_venv_installed(tmp_path_factory):
         assert subprocess.check_call(shlex.split("poetry install --no-interaction --no-root")) == 0
     return tmp_path_proj
 
-
-@pytest.fixture(scope="session")
-def requirements_txt_dir_with_venv_installed(tmp_path_factory):
-    tmp_path_proj = tmp_path_factory.getbasetemp() / "project_with_requirements_txt"
-    shutil.copytree("tests/data/project_with_requirements_txt", str(tmp_path_proj))
-    with run_within_dir(tmp_path_proj):
-        assert (
-            subprocess.check_call(
-                shlex.split(
-                    "pip install -r requirements.txt -r requirements-dev.txt -r requirements-2.txt -r"
-                    " requirements-typing.txt"
-                )
-            )
-            == 0
-        )
-    return tmp_path_proj
-
-
 def test_cli_returns_error(dir_with_venv_installed):
     with run_within_dir(dir_with_venv_installed):
         result = subprocess.run(shlex.split("poetry run deptry ."), capture_output=True, text=True)
@@ -88,51 +70,6 @@ def test_cli_extend_exclude(dir_with_venv_installed):
         )
         assert result.returncode == 1
         assert "The project contains obsolete dependencies:\n\n\tisort\n\trequests\n\ttoml\n\n" in result.stderr
-
-
-def test_cli_single_requirements_txt(requirements_txt_dir_with_venv_installed):
-    with run_within_dir(requirements_txt_dir_with_venv_installed):
-        result = subprocess.run(
-            shlex.split("deptry . --requirements-txt requirements.txt --requirements-txt-dev requirements-dev.txt"),
-            capture_output=True,
-            text=True,
-        )
-        assert result.returncode == 1
-        assert (
-            "The project contains obsolete dependencies:\n\n\tisort\n\trequests\n\nConsider removing them from your"
-            " project's dependencies. If a package is used for development purposes, you should add it to your"
-            " development dependencies instead.\n\n-----------------------------------------------------\n\nThere are"
-            " dependencies missing from the project's list of dependencies:\n\n\twhite\n\nConsider adding them to your"
-            " project's dependencies. \n\n-----------------------------------------------------\n\nThere are transitive"
-            " dependencies that should be explicitly defined as dependencies:\n\n\turllib3\n\nThey are currently"
-            " imported but not specified directly as your project's"
-            " dependencies.\n\n-----------------------------------------------------\n\nThere are imported modules from"
-            " development dependencies detected:\n\n\tblack\n\n"
-            in result.stderr
-        )
-
-
-def test_cli_multiple_requirements_txt(requirements_txt_dir_with_venv_installed):
-    with run_within_dir(requirements_txt_dir_with_venv_installed):
-        result = subprocess.run(
-            shlex.split(
-                "deptry . --requirements-txt requirements.txt,requirements-2.txt --requirements-txt-dev"
-                " requirements-dev.txt,requirements-typing.txt"
-            ),
-            capture_output=True,
-            text=True,
-        )
-        assert result.returncode == 1
-        assert (
-            "The project contains obsolete dependencies:\n\n\tisort\n\trequests\n\nConsider removing them from your"
-            " project's dependencies. If a package is used for development purposes, you should add it to your"
-            " development dependencies instead.\n\n-----------------------------------------------------\n\nThere are"
-            " dependencies missing from the project's list of dependencies:\n\n\twhite\n\nConsider adding them to your"
-            " project's dependencies. \n\n-----------------------------------------------------\n\nThere are imported"
-            " modules from development dependencies detected:\n\n\tblack\n\n"
-            in result.stderr
-        )
-
 
 def test_cli_verbose(dir_with_venv_installed):
     with run_within_dir(dir_with_venv_installed):
