+++ tests/data/project_with_pdm/README.md
@@ -1,0 +1,54 @@
+## Dependencies
+
+Dependencies are:
+
+```
+toml
+urllib3
+isort
+click
+requests
+pkginfo
+```
+
+dev-dependencies:
+
+```
+black
+```
+
+## Imports
+
+Imported in .py files are
+
+```
+click
+urllib3
+black
+white
+```
+
+Additional imports in .ipynb file:
+
+```
+toml
+```
+
+## Config
+
+pyproject.toml specifies to ignore the dependency:
+
+```
+pkginfo
+```
+
+## Output
+
+So expected output without any additional configuration:
+
+```
+obsolete: requests (because pkginfo is ignored)
+missing: white
+transitive: None
+dev: black
+```
