+++ tests/data/project_with_pdm/pyproject.toml
@@ -1,0 +1,25 @@
+[project]
+# PEP 621 project metadata
+# See https://www.python.org/dev/peps/pep-0621/
+requires-python = ">=3.7"
+dependencies = [
+    "toml",
+    "urllib3>=1.26.12",
+    "isort>=5.10.1",
+    "click>=8.1.3",
+    "requests>=2.28.1",
+    "pkginfo>=1.8.3",
+]
+
+[tool.pdm]
+version = {source = "scm"}
+
+[tool.pdm.dev-dependencies]
+lint = [
+    "black>=22.6.0"
+]
+
+[tool.deptry]
+ignore_obsolete = [
+'pkginfo'
+]
