+++ tests/data/project_with_pdm/src/main.py
@@ -1,0 +1,7 @@
+from os import chdir, walk
+from pathlib import Path
+
+import black
+import click
+import white as w
+from urllib3 import contrib
